# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AgeConfirmationRequest": {
        "properties": {
          "confirmed": {
            "const": true,
            "title": "Confirmed",
            "type": "boolean"
          }
        },
        "required": [
          "confirmed"
        ],
        "title": "AgeConfirmationRequest",
        "type": "object"
      },
      "AgeConfirmationResponse": {
        "properties": {
          "ageConfirmed": {
            "title": "Ageconfirmed",
            "type": "boolean"
          },
          "confirmedAt": {
            "format": "date-time",
            "title": "Confirmedat",
            "type": "string"
          }
        },
        "required": [
          "ageConfirmed",
          "confirmedAt"
        ],
        "title": "AgeConfirmationResponse",
        "type": "object"
      },
      "AiAllowance": {
        "properties": {
          "dailyMessageLimit": {
            "title": "Dailymessagelimit",
            "type": "integer"
          },
          "messagesUsedToday": {
            "title": "Messagesusedtoday",
            "type": "integer"
          }
        },
        "required": [
          "dailyMessageLimit",
          "messagesUsedToday"
        ],
        "title": "AiAllowance",
        "type": "object"
      },
...
      "BootstrapResponse": {
        "properties": {
          "accountSecurity": {
            "$ref": "#/components/schemas/AccountSecurityResponse"
          },
          "ai": {
            "$ref": "#/components/schemas/AiAllowance"
          },
          "cellar": {
            "$ref": "#/components/schemas/CellarListResponse"
          },
          "featureFlags": {
            "$ref": "#/components/schemas/FeatureFlags"
          },
          "privacy": {
            "$ref": "#/components/schemas/PrivacySettingsResponse"
          },
          "profile": {
            "$ref": "#/components/schemas/UserProfileResponse"
          },
          "user": {
            "$ref": "#/components/schemas/AuthenticatedUser"
          }
        },
        "required": [
          "user",
          "profile",
          "privacy",
          "accountSecurity",
          "cellar",
          "ai",
          "featureFlags"
        ],
        "title": "BootstrapResponse",
        "type": "object"
      },
      "CellarBatchRequest": {
        "properties": {
          "ingredientIds": {
            "items": {
              "type": "string"
            },
            "maxItems": 200,
            "title": "Ingredientids",
            "type": "array"
          }
        },
        "required": [
          "ingredientIds"
        ],
        "title": "CellarBatchRequest",
        "type": "object"
      },
      "CellarItemCreate": {
        "properties": {
          "amountLabel": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amountlabel"
          },
          "customName": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customname"
          },
          "ingredientId": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingredientid"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "title": "CellarItemCreate",
        "type": "object"
      },
      "CellarItemPatch": {
        "properties": {
          "amountLabel": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amountlabel"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "title": "CellarItemPatch",
        "type": "object"
      },
      "CellarItemResponse": {
        "properties": {
          "amountLabel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amountlabel"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "customName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customname"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "ingredientId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingredientid"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "source": {
            "$ref": "#/components/schemas/CellarItemSource"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "ingredientId",
          "customName",
          "amountLabel",
          "note",
          "source",
          "createdAt",
          "updatedAt"
        ],
        "title": "CellarItemResponse",
        "type": "object"
      },
      "CellarItemSource": {
        "enum": [
          "MANUAL",
          "LOCAL_SYNC"
        ],
        "title": "CellarItemSource",
        "type": "string"
      },
      "CellarListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CellarItemResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "CellarListResponse",
        "type": "object"
      },
      "DeleteAccountRequest": {
        "properties": {
          "confirmation": {
            "const": "DELETE",
            "title": "Confirmation",
            "type": "string"
          }
        },
        "required": [
          "confirmation"
        ],
        "title": "DeleteAccountRequest",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorPayload"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "ErrorPayload": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details"
        ],
        "title": "ErrorPayload",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LocalSyncRequest": {
        "properties": {
          "ageVerified": {
            "default": false,
            "title": "Ageverified",
            "type": "boolean"
          },
          "cellarIngredientIds": {
            "items": {
              "type": "string"
            },
            "maxItems": 200,
            "title": "Cellaringredientids",
            "type": "array"
          },
          "privacySettings": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PrivacySettingsPatch"
              },
              {
                "type": "null"
              }
            ]
          },
          "profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserProfilePatch"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "LocalSyncRequest",
        "type": "object"
      },
...
      "PrivacySettingsPatch": {
        "properties": {
          "analyticsOptIn": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyticsoptin"
          },
          "localOnlyMode": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Localonlymode"
          },
          "syncWhenLoggedIn": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Syncwhenloggedin"
          }
        },
        "title": "PrivacySettingsPatch",
        "type": "object"
      },
      "PrivacySettingsResponse": {
        "properties": {
          "analyticsOptIn": {
            "title": "Analyticsoptin",
            "type": "boolean"
          },
          "localOnlyMode": {
            "title": "Localonlymode",
            "type": "boolean"
          },
          "syncWhenLoggedIn": {
            "title": "Syncwhenloggedin",
            "type": "boolean"
          }
        },
        "required": [
          "localOnlyMode",
          "analyticsOptIn",
          "syncWhenLoggedIn"
        ],
        "title": "PrivacySettingsResponse",
        "type": "object"
      },
...
      "UserProfilePatch": {
        "properties": {
          "avatarKey": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatarkey"
          },
          "birthday": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birthday"
          },
          "city": {
            "anyOf": [
              {
                "maxLength": 12,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "gender": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "nickname": {
            "anyOf": [
              {
                "maxLength": 16,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          },
          "occupation": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occupation"
          },
          "school": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "School"
          },
          "showAge": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Showage"
          },
          "showBirthdayTag": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Showbirthdaytag"
          },
          "showZodiac": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Showzodiac"
          },
          "signature": {
            "anyOf": [
              {
                "maxLength": 60,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signature"
          }
        },
        "title": "UserProfilePatch",
        "type": "object"
      },
      "UserProfileResponse": {
        "properties": {
          "avatarKey": {
            "title": "Avatarkey",
            "type": "string"
          },
          "avatarUri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avataruri"
          },
          "birthday": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birthday"
          },
          "city": {
            "title": "City",
            "type": "string"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "nickname": {
            "title": "Nickname",
            "type": "string"
          },
          "occupation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occupation"
          },
          "school": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "School"
          },
          "showAge": {
            "title": "Showage",
            "type": "boolean"
          },
          "showBirthdayTag": {
            "title": "Showbirthdaytag",
            "type": "boolean"
          },
          "showZodiac": {
            "title": "Showzodiac",
            "type": "boolean"
          },
          "signature": {
            "title": "Signature",
            "type": "string"
          }
        },
        "required": [
          "nickname",
          "avatarKey",
          "signature",
          "city",
          "gender",
          "birthday",
          "showBirthdayTag",
          "showAge",
          "showZodiac",
          "occupation",
          "school"
        ],
        "title": "UserProfileResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cellar/items": {
      "get": {
        "operationId": "items_api_v1_cellar_items_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CellarListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Items",
        "tags": [
          "api",
          "cellar"
        ]
      },
      "post": {
        "operationId": "create_item_api_v1_cellar_items_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CellarItemCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CellarItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Item",
        "tags": [
          "api",
          "cellar"
        ]
      }
    },
    "/api/v1/cellar/items/batch": {
      "post": {
        "operationId": "batch_items_api_v1_cellar_items_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CellarBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CellarListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Batch Items",
        "tags": [
          "api",
          "cellar"
        ]
      }
    },
    "/api/v1/cellar/items/{item_id}": {
      "delete": {
        "operationId": "remove_item_api_v1_cellar_items__item_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove Item",
        "tags": [
          "api",
          "cellar"
        ]
      },
      "patch": {
        "operationId": "patch_item_api_v1_cellar_items__item_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CellarItemPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CellarItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Patch Item",
        "tags": [
          "api",
          "cellar"
        ]
      }
    },
    "/api/v1/me/account": {
      "delete": {
        "operationId": "remove_account_api_v1_me_account_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove Account",
        "tags": [
          "api",
          "me"
        ]
      }
    },
    "/api/v1/me/age-confirmation": {
      "post": {
        "operationId": "age_confirmation_api_v1_me_age_confirmation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgeConfirmationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgeConfirmationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Age Confirmation",
        "tags": [
          "api",
          "me"
        ]
      }
    },
    "/api/v1/me/bootstrap": {
      "get": {
        "operationId": "bootstrap_api_v1_me_bootstrap_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BootstrapResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Bootstrap",
        "tags": [
          "api",
          "me"
        ]
      }
    },
    "/api/v1/me/local-sync": {
      "post": {
        "operationId": "local_sync_api_v1_me_local_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalSyncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BootstrapResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Local Sync",
        "tags": [
          "api",
          "me"
        ]
      }
    },
    "/api/v1/me/privacy": {
      "patch": {
        "operationId": "patch_privacy_api_v1_me_privacy_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrivacySettingsPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacySettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Patch Privacy",
        "tags": [
          "api",
          "me"
        ]
      }
    },
    "/api/v1/me/profile": {
      "get": {
        "operationId": "get_profile_api_v1_me_profile_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Profile",
        "tags": [
          "api",
          "me"
        ]
      },
      "patch": {
        "operationId": "patch_profile_api_v1_me_profile_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfilePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Patch Profile",
        "tags": [
          "api",
          "me"
        ]
      }
    },
...
  }
}